+++ tests/test_component_set/components/aggregator/components/nested_hello/include/nested_hello.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+int nested_hello();