+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ## [Unreleased]
 
-- ...
+### Added
+
+- `add_component_set` to register a group of components without creating local targets.
 
 ## [6.3.0] - 2025-11-18